+++ packages/std/src/transactions.rs
@@ -39,12 +39,12 @@ impl<'a, S: ReadonlyStorage> StorageTransaction<'a, S> {
 }
 
 impl<'a, S: ReadonlyStorage> ReadonlyStorage for StorageTransaction<'a, S> {
-    fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
+    fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
         match self.local_state.get(key) {
-            Some(val) => match val {
+            Some(val) => Ok(match val {
                 Delta::Set { value } => Some(value.clone()),
                 Delta::Delete {} => None,
-            },
+            }),
             None => self.storage.get(key),
         }
     }
@@ -276,13 +276,13 @@ mod test {
         check.set(b"food", b"bank").unwrap();
         check.remove(b"foo").unwrap();
 
-        assert_eq!(None, check.get(b"foo"));
-        assert_eq!(Some(b"bank".to_vec()), check.get(b"food"));
+        assert_eq!(None, check.get(b"foo").unwrap());
+        assert_eq!(Some(b"bank".to_vec()), check.get(b"food").unwrap());
 
         // now commit to base and query there
         check.prepare().commit(&mut base).unwrap();
-        assert_eq!(None, base.get(b"foo"));
-        assert_eq!(Some(b"bank".to_vec()), base.get(b"food"));
+        assert_eq!(None, base.get(b"foo").unwrap());
+        assert_eq!(Some(b"bank".to_vec()), base.get(b"food").unwrap());
     }
 
     #[test]
@@ -293,13 +293,13 @@ mod test {
         check.set(b"food", b"bank").unwrap();
         check.remove(b"foo").unwrap();
 
-        assert_eq!(None, check.get(b"foo"));
-        assert_eq!(Some(b"bank".to_vec()), check.get(b"food"));
+        assert_eq!(None, check.get(b"foo").unwrap());
+        assert_eq!(Some(b"bank".to_vec()), check.get(b"food").unwrap());
 
         // now commit to base and query there
         check.prepare().commit(&mut base).unwrap();
-        assert_eq!(None, base.get(b"foo"));
-        assert_eq!(Some(b"bank".to_vec()), base.get(b"food"));
+        assert_eq!(None, base.get(b"foo").unwrap());
+        assert_eq!(Some(b"bank".to_vec()), base.get(b"food").unwrap());
     }
 
     #[test]
@@ -337,11 +337,11 @@ mod test {
         base.set(b"foo", b"bar").unwrap();
 
         let mut check = StorageTransaction::new(&base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
+        assert_eq!(check.get(b"foo").unwrap(), Some(b"bar".to_vec()));
         check.set(b"subtx", b"works").unwrap();
         check.prepare().commit(&mut base).unwrap();
 
-        assert_eq!(base.get(b"subtx"), Some(b"works".to_vec()));
+        assert_eq!(base.get(b"subtx").unwrap(), Some(b"works".to_vec()));
     }
 
     #[test]
@@ -351,16 +351,16 @@ mod test {
 
         let mut stxn1 = StorageTransaction::new(&base);
 
-        assert_eq!(stxn1.get(b"foo"), Some(b"bar".to_vec()));
+        assert_eq!(stxn1.get(b"foo").unwrap(), Some(b"bar".to_vec()));
 
         stxn1.set(b"subtx", b"works").unwrap();
-        assert_eq!(stxn1.get(b"subtx"), Some(b"works".to_vec()));
+        assert_eq!(stxn1.get(b"subtx").unwrap(), Some(b"works".to_vec()));
 
         // Can still read from base, txn is not yet committed
-        assert_eq!(base.get(b"subtx"), None);
+        assert_eq!(base.get(b"subtx").unwrap(), None);
 
         stxn1.prepare().commit(&mut base).unwrap();
-        assert_eq!(base.get(b"subtx"), Some(b"works".to_vec()));
+        assert_eq!(base.get(b"subtx").unwrap(), Some(b"works".to_vec()));
     }
 
     #[test]
@@ -369,11 +369,11 @@ mod test {
         base.set(b"foo", b"bar").unwrap();
 
         let mut check = StorageTransaction::new(&base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
+        assert_eq!(check.get(b"foo").unwrap(), Some(b"bar".to_vec()));
         check.set(b"subtx", b"works").unwrap();
         check.rollback();
 
-        assert_eq!(base.get(b"subtx"), None);
+        assert_eq!(base.get(b"subtx").unwrap(), None);
     }
 
     #[test]
@@ -382,10 +382,10 @@ mod test {
         base.set(b"foo", b"bar").unwrap();
 
         let mut check = StorageTransaction::new(&base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
+        assert_eq!(check.get(b"foo").unwrap(), Some(b"bar".to_vec()));
         check.set(b"subtx", b"works").unwrap();
 
-        assert_eq!(base.get(b"subtx"), None);
+        assert_eq!(base.get(b"subtx").unwrap(), None);
     }
 
     #[test]
@@ -396,24 +396,24 @@ mod test {
         // writes on success
         let res: Result<i32> = transactional(&mut base, &|store| {
             // ensure we can read from the backing store
-            assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
+            assert_eq!(store.get(b"foo").unwrap(), Some(b"bar".to_vec()));
             // we write in the Ok case
             store.set(b"good", b"one").unwrap();
             Ok(5)
         });
         assert_eq!(5, res.unwrap());
-        assert_eq!(base.get(b"good"), Some(b"one".to_vec()));
+        assert_eq!(base.get(b"good").unwrap(), Some(b"one".to_vec()));
 
         // rejects on error
         let res: Result<i32> = transactional(&mut base, &|store| {
             // ensure we can read from the backing store
-            assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
-            assert_eq!(store.get(b"good"), Some(b"one".to_vec()));
+            assert_eq!(store.get(b"foo").unwrap(), Some(b"bar".to_vec()));
+            assert_eq!(store.get(b"good").unwrap(), Some(b"one".to_vec()));
             // we write in the Error case
             store.set(b"bad", b"value").unwrap();
             Unauthorized.fail()
         });
         assert!(res.is_err());
-        assert_eq!(base.get(b"bad"), None);
+        assert_eq!(base.get(b"bad").unwrap(), None);
     }
 }